Inject backend URI through environmental vars
Previously config.js used to resolve the backend URI depending on
window.location. This was a flaky soluion as the URI can really be anything.

The URI is now injected from an environmental variable at build time. This also
required a change in the Dockerfile, as the production build was previously
built when the image was built. It is now built when the image starts.

docker-compose.yml needs an addition due to this:

services:
  # ...
  environment:
    # uri to the backend, different on dev, staging and prod
    - 'REACT_APP_CLIENTSIDE_BACKEND_URI=http://localhost:7001'