Reorder email config "updating templates" asserts
Reordered tests so that cypress will first assert that the target template has
been updated, before checking that the other text fields are empty. If Cypress
checks that the other fields are empty before doing the actual check, it's
possible that the actual values from the backend haven't propagated all the way
to the rendered state meaning that the empty checks pass before the actual
values are shown.